+++ tests/test_parser.ads
@@ -1,7 +1,0 @@
---  Parser tests for TinyAML
-
-package Test_Parser is
-
-   procedure Run_Tests;
-
-end Test_Parser;